Update and rename spinning_cube.c to spinning_oct.c
Chenced all the 'cube' names to 'oct' names.
